requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
firebase-admin==6.0.1
googlemaps==4.10.0
google-cloud-aiplatform==1.30.0
requests==2.31.0
gunicorn==21.2.0